Tests/LibWeb/Text/expected/scroll-into-view-with-pinch-zoom.txt

target visible before scrollIntoView: false
target visible after scrollIntoView: true
